requirements.txt
# Flask
Flask==3.0.0
flask-cors==4.0.0

# Google Cloud
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1
vertexai==1.38.1

# Graph
networkx==3.2.1

# HTTP & API
requests==2.31.0

# Utilities
python-dotenv==1.0.0
